Clean up analysis files
