--- papylio/analysis/decorators.py
@@ -1,44 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Sun Jun  2 01:48:30 2019
-
-@author: iason
-
-Usefull decorators
-"""
-
-import functools
-import time
-
-
-# General decorator template
-def decorator(func):
-    @functools.wraps(func)
-    def wrapper_decorator(*args, **kwargs):
-        # Do something before
-        value = func(*args, **kwargs)
-        # Do something after
-        return value
-    return wrapper_decorator
-
-
-
-def timer(func):
-    """Print the runtime of the decorated function"""
-    @functools.wraps(func)
-    def wrapper_timer(*args, **kwargs):
-        start_time = time.perf_counter()    # 1
-        value = func(*args, **kwargs)
-        end_time = time.perf_counter()      # 2
-        run_time = end_time - start_time    # 3
-        print(f"Finished {func.__name__!r} in {run_time:.4f} sec")
-        return value
-    return wrapper_timer
-
-@timer
-def waste_some_time(num_times):
-    for _ in range(num_times):
-        sum([i**2 for i in range(10000)])
-
-if __name__ == '__main__':
-    waste_some_time(100)
