+++ papylio/timer.py
@@ -1,4 +1,5 @@
 import time
+import functools
 
 class TimerError(Exception):
     """A custom exception used to report errors in use of Timer class"""
@@ -30,3 +31,16 @@ class Timer:
     def startstop(self, name=''):
         self.stop()
         self.start(name=name)
+
+# Timer decorator
+def timer(func):
+    """Print the runtime of the decorated function"""
+    @functools.wraps(func)
+    def wrapper_timer(*args, **kwargs):
+        start_time = time.perf_counter()    # 1
+        value = func(*args, **kwargs)
+        end_time = time.perf_counter()      # 2
+        run_time = end_time - start_time    # 3
+        print(f"Finished {func.__name__!r} in {run_time:.4f} sec")
+        return value
+    return wrapper_timer
